@@ -77,62 +77,62 @@
     ]
   },
   "devDependencies": {
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^11.0.1",
-    "@rollup/plugin-replace": "^2.3.4",
-    "@types/jest": "^26.0.19",
-    "@types/lodash.camelcase": "^4.3.6",
-    "@types/lodash.get": "^4.4.6",
+    "@commitlint/cli": "^17.4.4",
+    "@commitlint/config-conventional": "^17.4.4",
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-multi-entry": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-replace": "^5.0.2",
+    "@rollup/plugin-terser": "^0.4.0",
+    "@types/jest": "^29.4.0",
+    "@types/lodash.camelcase": "^4.3.7",
+    "@types/lodash.get": "^4.4.7",
     "@types/nanoevents": "^1.0.0",
-    "@types/node": "^14.14.16",
-    "@types/rollup-plugin-json": "^3.0.2",
+    "@types/node": "^18.14.2",
+    "@types/rollup-plugin-json": "^3.0.3",
     "@types/rollup-plugin-sourcemaps": "^0.4.2",
-    "@types/three": "^0.132.0",
-    "@typescript-eslint/eslint-plugin": "^4.32.0",
-    "@typescript-eslint/parser": "^4.32.0",
+    "@types/three": "^0.149.0",
+    "@typescript-eslint/eslint-plugin": "^5.54.0",
+    "@typescript-eslint/parser": "^5.54.0",
     "colors": "^1.4.0",
-    "commitizen": "^4.2.2",
-    "coveralls": "^3.1.0",
-    "cpy-cli": "^3.1.1",
+    "commitizen": "^4.3.0",
+    "coveralls": "^3.1.1",
+    "cpy-cli": "^4.2.0",
     "cross-env": "^7.0.3",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "husky": "^4.3.6",
-    "jest": "^26.6.3",
-    "jest-config": "^26.6.3",
-    "lint-staged": "^10.5.3",
-    "prettier": "^2.2.1",
-    "prompt": "^1.1.0",
-    "replace-in-file": "^6.1.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.35.1",
+    "eslint": "^8.35.0",
+    "eslint-config-prettier": "^8.6.0",
+    "husky": "^8.0.3",
+    "jest": "^29.4.3",
+    "jest-config": "^29.4.3",
+    "lint-staged": "^13.1.2",
+    "prettier": "^2.8.4",
+    "prompt": "^1.3.0",
+    "replace-in-file": "^6.3.5",
+    "rimraf": "^4.1.2",
+    "rollup": "^3.17.3",
     "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-multi-input": "^1.3.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-serve": "^1.1.0",
+    "rollup-plugin-serve": "^2.0.2",
     "rollup-plugin-sourcemaps": "^0.6.3",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.29.0",
-    "shelljs": "^0.8.4",
-    "three": "^0.132.2",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "shelljs": "^0.8.5",
+    "three": "^0.150.1",
     "three-ik": "^0.1.0",
-    "three-spritetext": "^1.6.2",
+    "three-spritetext": "^1.7.1",
     "travis-deploy-once": "^5.0.11",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^9.1.1",
-    "typedoc": "^0.19.2",
+    "ts-jest": "^29.0.5",
+    "ts-node": "^10.9.1",
+    "typedoc": "^0.23.26",
     "typedoc-plugin-internal-external": "^2.2.0",
-    "typedoc-plugin-no-inherit": "^1.2.0",
-    "typescript": "^4.1.3"
+    "typedoc-plugin-no-inherit": "^1.4.0",
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "three": ">=0.132.2"
   },
   "dependencies": {
-    "nanoevents": "^2.0.0"
+    "nanoevents": "^7.0.1"
   }
 }
